pyproject.toml
[project]
name = "foothill-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "mcp>=1.2.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
